+++ tests/plugins/clipboard/_helpers/pasting.js
@@ -15,18 +15,8 @@ function assertPasteEvent( editor, eventData, expected, message, async, skipCanc
 		delete expected.priority;
 	}
 
-	// Type doesn't have to be specified.
-	if ( !eventData.type )
-		eventData.type = 'auto';
-
-	eventData.method = 'paste';
-	// Allow passing a dataTransfer mock.
-	if ( !eventData.dataTransfer ) {
-		eventData.dataTransfer = new CKEDITOR.plugins.clipboard.dataTransfer();
-	}
-
 	editor.once( 'paste', onPaste, null, null, priority );
-	editor.fire( 'paste', eventData );
+	paste( editor, eventData );
 
 	if ( async )
 		wait();
@@ -60,6 +50,21 @@ function assertPasteEvent( editor, eventData, expected, message, async, skipCanc
 		else
 			assertPaste( data );
 	}
+}
+
+function paste( editor, eventData ) {
+	// Type doesn't have to be specified.
+	if ( !eventData.type ) {
+		eventData.type = 'auto';
+	}
+
+	eventData.method = 'paste';
+	// Allow passing a dataTransfer mock.
+	if ( !eventData.dataTransfer ) {
+		eventData.dataTransfer = new CKEDITOR.plugins.clipboard.dataTransfer();
+	}
+
+	editor.fire( 'paste', eventData );
 }
 
 function pasteFiles( editor, files, dataValue, pasteData ) {