+++ tests/plugins/pastetools/_helpers/ptTools.js
@@ -1,0 +1,35 @@
+/* exported pfwTools */
+
+( function() {
+	'use strict';
+
+	window.ptTools = {
+		ignoreTestsOnMobiles: function( tests ) {
+			tests._should = tests._should || { ignore: {} };
+
+			if ( bender.tools.env.mobile ) {
+				CKEDITOR.tools.array.forEach( CKEDITOR.tools.object.keys( tests ), function( tcName ) {
+					tests._should.ignore[ tcName ] = true;
+				} );
+			}
+
+			return tests;
+		},
+
+		loadFilters: function loadFilters( filters, callback ) {
+			var loaded = 0,
+				i;
+
+			for ( i = 0; i < filters.length; i++ ) {
+				( function( current ) {
+					CKEDITOR.scriptLoader.queue( current, function() {
+						if ( ++loaded === filters.length ) {
+							callback();
+						}
+					} );
+				}( filters[ i ] ) );
+			}
+		}
+	};
+} )();
+